+++ Wire-iOS/Sources/UserInterface/Conversation/Content/AVPlayerViewController+StatusBarAndNotification.swift
@@ -1,0 +1,38 @@
+//
+// Wire
+// Copyright (C) 2018 Wire Swiss GmbH
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program. If not, see http://www.gnu.org/licenses/.
+//
+
+import Foundation
+
+extension AVPlayerViewController {
+    override open var prefersStatusBarHidden: Bool {
+        get {
+            return true;
+        }
+    }
+
+    open override func viewDidDisappear(_ animated: Bool) {
+        super.viewDidDisappear(animated)
+        guard self.isBeingDismissed else {
+            return
+        }
+
+        ///TODO: notification name
+        NotificationCenter.default.post(name: NSNotification.Name(rawValue: "DismissingAVPlayer"), object: self)
+    }
+}
+
